FEAT(#24): RAG와 LLM 연결 및 API 구현

+++ app/domain/portfolio/schema/portfolioResult.py
@@ -11,7 +11,7 @@ class CompanyResult(CamelModel):
     company_name: str
     corp_code: str
     allocated_amount: int
-    final_score: int
+    final_score: float
     rank_no: int
     per: float
     roe: float
